# apps/web/openapi.json
{
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/movies/{id}": {
			"get": {
				"operationId": "MoviesController_findOne_v1",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Movie ID",
						"schema": { "type": "number" }
					}
				],
				"responses": {
					"200": {
						"description": "Movie found",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/MovieResponseDto" } }
						}
					},
					"404": { "description": "Movie not found" }
				},
				"security": [{ "access-token": [] }],
				"summary": "Get movie by id",
				"tags": ["Movies"]
			},
			"patch": {
				"operationId": "MoviesController_update_v1",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Movie ID",
						"schema": { "type": "number" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/MovieUpdateDto" } }
					}
				},
				"responses": {
					"200": {
						"description": "Movie updated",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/MovieResponseDto" } }
						}
					},
					"403": { "description": "Forbidden - Admin only" },
					"404": { "description": "Movie not found" }
				},
				"security": [{ "access-token": [] }],
				"summary": "Update movie (Admin only)",
				"tags": ["Movies"]
			},
			"delete": {
				"operationId": "MoviesController_remove_v1",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Movie ID",
						"schema": { "type": "number" }
					}
				],
				"responses": {
					"204": { "description": "Movie deleted" },
					"403": { "description": "Forbidden - Admin only" },
					"404": { "description": "Movie not found" }
				},
				"security": [{ "access-token": [] }],
				"summary": "Delete movie (Admin only)",
				"tags": ["Movies"]
			}
		},
...
		"/api/v1/groups/{id}": {
			"get": {
				"operationId": "GroupsController_findOne_v1",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Group ID",
						"schema": { "type": "number" }
					}
				],
				"responses": {
					"200": {
						"description": "Group found",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/GroupResponseDto" } }
						}
					},
					"403": { "description": "Forbidden - Not a group member" },
					"404": { "description": "Group not found" }
				},
				"security": [{ "access-token": [] }],
				"summary": "Get group by id (Group members only)",
				"tags": ["Groups"]
			},
			"patch": {
				"operationId": "GroupsController_update_v1",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Group ID",
						"schema": { "type": "number" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/GroupUpdateDto" } }
					}
				},
				"responses": {
					"200": {
						"description": "Group updated",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/GroupResponseDto" } }
						}
					},
					"403": { "description": "Forbidden - Insufficient permissions" },
					"404": { "description": "Group not found" }
				},
				"security": [{ "access-token": [] }],
				"summary": "Update group (Group moderators only)",
				"tags": ["Groups"]
			},
			"delete": {
				"operationId": "GroupsController_remove_v1",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Group ID",
						"schema": { "type": "number" }
					}
				],
				"responses": {
					"204": { "description": "Group deleted" },
					"403": { "description": "Forbidden - Not the group admin" },
					"404": { "description": "Group not found" }
				},
				"security": [{ "access-token": [] }],
				"summary": "Delete group (Group admin only)",
				"tags": ["Groups"]
			}
		},
...
		"/api/v1/groups/{groupId}/movies": {
			"post": {
				"operationId": "GroupMoviesController_addProviderMovie_v1",
				"parameters": [
					{
						"name": "groupId",
						"required": true,
						"in": "path",
						"description": "Group ID",
						"schema": { "type": "number" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/AddMovieDto" } }
					}
				},
				"responses": {
					"201": {
						"description": "Movie added to group",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/GroupMovieResponseDto" }
							}
						}
					},
					"403": { "description": "Forbidden - Not a group moderator" },
					"409": { "description": "Movie already added to this group" }
				},
				"security": [{ "access-token": [] }],
				"summary": "Add provider movie to group (Group moderators only)",
				"tags": ["Groups / Movies"]
			},
			"get": {
				"operationId": "GroupMovieDetailsController_findAll_v1",
				"parameters": [
					{
						"name": "groupId",
						"required": true,
						"in": "path",
						"description": "Group ID",
						"schema": { "type": "number" }
					},
					{
						"name": "status",
						"required": false,
						"in": "query",
						"description": "Filter by movie status",
						"schema": { "type": "string", "enum": ["tracking", "planned", "watched"] }
					},
					{
						"name": "query",
						"required": false,
						"in": "query",
						"description": "Search query for movie title",
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "List of group movies",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": { "$ref": "#/components/schemas/GroupMovieResponseDto" }
								}
							}
						}
					},
					"400": { "description": "Invalid query parameters" },
					"403": { "description": "Forbidden - Not a group member" }
				},
				"security": [{ "access-token": [] }],
				"summary": "Get all movies in group (Group members only)",
				"tags": ["Groups / Movies"]
			}
		},
		"/api/v1/groups/{groupId}/movies/custom": {
			"post": {
				"operationId": "GroupMoviesController_createCustomMovie_v1",
				"parameters": [
					{
						"name": "groupId",
						"required": true,
						"in": "path",
						"description": "Group ID",
						"schema": { "type": "number" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/CreateCustomMovieDto" }
						}
					}
				},
				"responses": {
					"201": {
						"description": "Custom movie created",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/GroupMovieResponseDto" }
							}
						}
					},
					"400": { "description": "Invalid status/date combination" },
					"403": { "description": "Forbidden - Not a group moderator" }
				},
				"security": [{ "access-token": [] }],
				"summary": "Create custom movie in group (Group moderators only)",
				"tags": ["Groups / Movies"]
			}
		},
		"/api/v1/groups/{groupId}/movies/{id}": {
			"patch": {
				"operationId": "GroupMoviesController_update_v1",
				"parameters": [
					{
						"name": "groupId",
						"required": true,
						"in": "path",
						"description": "Group ID",
						"schema": { "type": "number" }
					},
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Group Movie ID",
						"schema": { "type": "number" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/GroupMovieUpdateDto" } }
					}
				},
				"responses": {
					"200": {
						"description": "Movie updated",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/GroupMovieResponseDto" }
							}
						}
					},
					"400": { "description": "Invalid status/date combination" },
					"403": { "description": "Forbidden - Not a group moderator" },
					"404": { "description": "Movie not found in group" }
				},
				"security": [{ "access-token": [] }],
				"summary": "Update movie in group (Group moderators only)",
				"tags": ["Groups / Movies"]
			},
			"delete": {
				"operationId": "GroupMoviesController_remove_v1",
				"parameters": [
					{
						"name": "groupId",
						"required": true,
						"in": "path",
						"description": "Group ID",
						"schema": { "type": "number" }
					},
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Group Movie ID",
						"schema": { "type": "number" }
					}
				],
				"responses": {
					"204": { "description": "Movie removed from group" },
					"403": { "description": "Forbidden - Not a group moderator" },
					"404": { "description": "Movie not found in group" }
				},
				"security": [{ "access-token": [] }],
				"summary": "Remove movie from group (Group moderators only)",
				"tags": ["Groups / Movies"]
			},
			"get": {
				"operationId": "GroupMovieDetailsController_findOne_v1",
				"parameters": [
					{
						"name": "groupId",
						"required": true,
						"in": "path",
						"description": "Group ID",
						"schema": { "type": "number" }
					},
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Group Movie ID",
						"schema": { "type": "number" }
					}
				],
				"responses": {
					"200": {
						"description": "Group movie found",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/GroupMovieResponseDto" }
							}
						}
					},
					"403": { "description": "Forbidden - Not a group member" },
					"404": { "description": "Movie not found in group" }
				},
				"security": [{ "access-token": [] }],
				"summary": "Get movie details in group (Group members only)",
				"tags": ["Groups / Movies"]
			}
		},
...
		"/api/v1/groups/{groupId}/movies/{groupMovieId}/reviews": {
			"post": {
				"operationId": "GroupMovieReviewsController_create_v1",
				"parameters": [
					{
						"name": "groupId",
						"required": true,
						"in": "path",
						"description": "Group ID",
						"schema": { "type": "number" }
					},
					{
						"name": "groupMovieId",
						"required": true,
						"in": "path",
						"description": "Group Movie ID",
						"schema": { "type": "number" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/CreateReviewDto" } }
					}
				},
				"responses": {
					"201": {
						"description": "Review created",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ReviewResponseDto" } }
						}
					},
					"400": { "description": "Invalid rating or text" },
					"403": { "description": "Forbidden - Not a group member" },
					"404": { "description": "Movie not found in group" },
					"409": { "description": "Review already exists" },
					"422": { "description": "Movie not watched" }
				},
				"security": [{ "access-token": [] }],
				"summary": "Create a review for a group movie (Group members only)",
				"tags": ["Groups / Movie Reviews"]
			}
		},
		"/api/v1/groups/{groupId}/movies/{groupMovieId}/reviews/{id}": {
			"patch": {
				"operationId": "GroupMovieReviewsController_update_v1",
				"parameters": [
					{
						"name": "groupId",
						"required": true,
						"in": "path",
						"description": "Group ID",
						"schema": { "type": "number" }
					},
					{
						"name": "groupMovieId",
						"required": true,
						"in": "path",
						"description": "Group Movie ID",
						"schema": { "type": "number" }
					},
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Review ID",
						"schema": { "type": "number" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/UpdateReviewDto" } }
					}
				},
				"responses": {
					"200": {
						"description": "Review updated",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ReviewResponseDto" } }
						}
					},
					"400": { "description": "Invalid request body" },
					"403": { "description": "Forbidden" },
					"404": { "description": "Review or movie not found" },
					"422": { "description": "Movie not watched" }
				},
				"security": [{ "access-token": [] }],
				"summary": "Update a review (Author or global admin only)",
				"tags": ["Groups / Movie Reviews"]
			},
			"delete": {
				"operationId": "GroupMovieReviewsController_delete_v1",
				"parameters": [
					{
						"name": "groupId",
						"required": true,
						"in": "path",
						"description": "Group ID",
						"schema": { "type": "number" }
					},
					{
						"name": "groupMovieId",
						"required": true,
						"in": "path",
						"description": "Group Movie ID",
						"schema": { "type": "number" }
					},
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Review ID",
						"schema": { "type": "number" }
					}
				],
				"responses": {
					"204": { "description": "Review deleted" },
					"403": { "description": "Forbidden" },
					"404": { "description": "Review or movie not found" }
				},
				"security": [{ "access-token": [] }],
				"summary": "Delete a review (Author or global admin only)",
				"tags": ["Groups / Movie Reviews"]
			}
		}
	},
	"info": {
		"title": "Movies App API",
		"description": "REST API for Movies application\n\n**Links:**\n- [OpenAPI JSON](/api/docs/json)\n- [OpenAPI YAML](/api/docs/yaml)",
		"version": "1.0.0",
		"contact": {}
	},
...
	"servers": [{ "url": "http://localhost:8080" }],
	"components": {
		"securitySchemes": {
			"access-token": {
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"type": "http",
				"description": "Enter JWT access token"
			}
		},
		"schemas": {
...
			"MovieResponseDto": {
				"type": "object",
				"properties": {
					"id": { "type": "number" },
					"externalId": { "type": "string" },
					"imdbId": { "type": "string", "nullable": true },
					"title": { "type": "string" },
					"posterPath": { "type": "string", "nullable": true },
					"overview": { "type": "string", "nullable": true },
					"releaseYear": { "type": "number", "nullable": true },
					"rating": { "type": "string", "nullable": true },
					"genres": { "nullable": true, "type": "array", "items": { "type": "string" } },
					"runtime": { "type": "number", "nullable": true },
					"createdAt": { "format": "date-time", "type": "string" },
					"updatedAt": { "format": "date-time", "type": "string" }
				},
				"required": [
					"id",
					"externalId",
					"imdbId",
					"title",
					"posterPath",
					"overview",
					"releaseYear",
					"rating",
					"genres",
					"runtime",
					"createdAt",
					"updatedAt"
				]
			},
...
			"MovieUpdateDto": {
				"type": "object",
				"properties": {
					"title": { "type": "string", "description": "Movie title" },
					"posterPath": { "type": "string", "description": "Poster path" },
					"overview": { "type": "string", "description": "Movie overview" },
					"releaseYear": { "type": "number", "description": "Release year", "example": 1999 },
					"rating": {
						"type": "number",
						"description": "Rating (0-10)",
						"minimum": 0,
						"maximum": 10
					},
					"runtime": { "type": "number", "description": "Runtime in minutes" }
				}
			},
			"GroupResponseDto": {
				"type": "object",
				"properties": {
					"id": { "type": "number" },
					"name": { "type": "string" },
					"description": { "type": "string", "nullable": true },
					"avatarUrl": { "type": "string", "nullable": true },
					"createdAt": { "format": "date-time", "type": "string" },
					"updatedAt": { "format": "date-time", "type": "string" },
					"currentUserRole": { "type": "string", "enum": ["admin", "moderator", "member"] }
				},
				"required": ["id", "name", "description", "avatarUrl", "createdAt", "updatedAt"]
			},
...
			"GroupUpdateDto": {
				"type": "object",
				"properties": {
					"name": { "type": "string", "example": "Movie Club" },
					"description": { "type": "string", "example": "Weekly movie nights" },
					"avatarUrl": { "type": "string", "example": "https://example.com/avatar.jpg" }
				}
			},
...
			"AddMovieDto": {
				"type": "object",
				"properties": {
					"imdbId": {
						"type": "string",
						"description": "IMDb ID (universal, takes priority)",
						"example": "tt0133093"
					},
					"externalId": {
						"type": "string",
						"description": "Provider external ID",
						"example": "301"
					}
				}
			},
			"ReviewResponseDto": {
				"type": "object",
				"properties": {
					"id": { "type": "number" },
					"groupMovieId": { "type": "number" },
					"userId": { "type": "number" },
					"userName": { "type": "string" },
					"rating": { "type": "number" },
					"text": { "type": "string", "nullable": true },
					"createdAt": { "format": "date-time", "type": "string" },
					"updatedAt": { "format": "date-time", "type": "string" },
					"isOwn": { "type": "boolean" }
				},
				"required": [
					"id",
					"groupMovieId",
					"userId",
					"userName",
					"rating",
					"text",
					"createdAt",
					"updatedAt"
				]
			},
			"GroupMovieResponseDto": {
				"type": "object",
				"properties": {
					"id": { "type": "number" },
					"groupId": { "type": "number" },
					"source": { "type": "string", "enum": ["provider", "custom"] },
					"movieId": { "type": "number", "nullable": true },
					"title": { "type": "string" },
					"posterPath": { "type": "string", "nullable": true },
					"overview": { "type": "string", "nullable": true },
					"releaseYear": { "type": "number", "nullable": true },
					"runtime": { "type": "number", "nullable": true },
					"rating": { "type": "string", "nullable": true },
					"addedBy": { "type": "number" },
					"status": { "type": "string", "enum": ["tracking", "planned", "watched"] },
					"watchDate": { "format": "date-time", "type": "string", "nullable": true },
					"createdAt": { "format": "date-time", "type": "string" },
					"updatedAt": { "format": "date-time", "type": "string" },
					"currentUserRole": { "type": "string", "enum": ["admin", "moderator", "member"] },
					"reviews": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/ReviewResponseDto" }
					},
					"averageRating": { "type": "number", "nullable": true },
					"reviewCount": { "type": "number" }
				},
				"required": [
					"id",
					"groupId",
					"source",
					"movieId",
					"title",
					"posterPath",
					"overview",
					"releaseYear",
					"runtime",
					"rating",
					"addedBy",
					"status",
					"watchDate",
					"createdAt",
					"updatedAt"
				]
			},
			"CreateCustomMovieDto": {
				"type": "object",
				"properties": {
					"title": { "type": "string", "description": "Movie title" },
					"posterPath": { "type": "string", "description": "Poster path" },
					"overview": { "type": "string", "description": "Movie overview" },
					"releaseYear": { "type": "number", "description": "Release year" },
					"runtime": { "type": "number", "description": "Runtime in minutes" },
					"status": {
						"type": "string",
						"description": "Movie status",
						"enum": ["tracking", "planned", "watched"],
						"default": "tracking"
					},
					"watchDate": {
						"type": "string",
						"nullable": true,
						"description": "Watch date (planned or actual, depending on status)",
						"example": "2024-12-31T20:00:00Z"
					}
				},
				"required": ["title"]
			},
			"GroupMovieUpdateDto": {
				"type": "object",
				"properties": {
					"status": {
						"type": "string",
						"description": "Movie status in group",
						"enum": ["tracking", "planned", "watched"]
					},
					"watchDate": {
						"type": "string",
						"nullable": true,
						"description": "Watch date (planned or actual, depending on status)",
						"example": "2024-12-31T20:00:00Z"
					},
					"title": { "type": "string", "description": "Movie title" },
					"posterPath": { "type": "string", "description": "Poster path" },
					"overview": { "type": "string", "description": "Movie overview" },
					"releaseYear": { "type": "number", "description": "Release year" },
					"runtime": { "type": "number", "description": "Runtime in minutes" }
				}
			},
...
			"CreateReviewDto": {
				"type": "object",
				"properties": {
					"rating": {
						"type": "number",
						"description": "Rating 0.5–5.0, step 0.5",
						"minimum": 0.5,
						"maximum": 5
					},
					"text": { "type": "string", "description": "Review text", "maxLength": 2000 }
				},
				"required": ["rating"]
			},
			"UpdateReviewDto": {
				"type": "object",
				"properties": {
					"rating": {
						"type": "number",
						"description": "Rating 0.5–5.0, step 0.5",
						"minimum": 0.5,
						"maximum": 5
					},
					"text": { "type": "string", "description": "Review text", "maxLength": 2000 }
				}
			}
		}
	},
	"security": [{ "access-token": [] }]
}